Fixed warning

+++ src/JSONSettings/Model/SettingDefinitionMgr.cpp
@@ -47,7 +47,7 @@ namespace systelab { namespace setting {
 										  const SettingDefinition& definition)
 	{
 		m_settings[file][sectionPath] = definition;
-		m_settings[file][sectionPath].id = m_settings[file].size();
+		m_settings[file][sectionPath].id = static_cast<int>(m_settings[file].size());
 	}
 
 	SecurityKey SettingDefinitionMgr::getSettingsFileEncryptionKey(const SettingsFile& file) const
